requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.3
Django==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
numpy==2.2.5
pandas==2.2.3
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
